% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_adjacency_matrix}
\alias{create_adjacency_matrix}
\title{Create adjacency matrix from V}
\usage{
create_adjacency_matrix(V, wt, n)
}
\arguments{
\item{V}{Matrix of centroid differences}

\item{wt}{Weights vector}

\item{n}{Number of observations in clustering}
}
\description{
\code{create_adjacency_matrix} creates an n-by-n adjacency matrix from the matrix of centroid differences.
}
