% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_E_matrix}
\alias{create_E_matrix}
\title{Create E matrix}
\usage{
create_E_matrix(n)
}
\arguments{
\item{n}{Number of data observations}
}
\value{
E The edge matrix.

nK The number of non-zero weights.
}
\description{
\code{create_E_matrix} generates the edges matrix where the edges are denoted 
as \code{l=(i,j)}.
}
\examples{
n <- 10
E <- create_E_matrix(n)$E
nK <- create_E_matrix(n)$nK
}
