% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Generate Distance Matrix from Data}
\usage{
distance_matrix(X)
}
\arguments{
\item{X}{data matrix}
}
\description{
\code{distance_matrix} generates the distance matrix from the data matrix.
}
