% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_clusters}
\alias{find_clusters}
\title{Find clusters}
\usage{
find_clusters(A)
}
\description{
\code{find_clusters} uses breadth-first search to identify the connected components
 of the corresponding adjacency graph of the centroid differences vectors.
 
 @param A adjacency matrix
 @import igraph
 @export
}
