% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_clusters_from_adjacency}
\alias{find_clusters_from_adjacency}
\title{Find clusters}
\usage{
find_clusters_from_adjacency(A)
}
\arguments{
\item{A}{adjacency matrix}
}
\description{
\code{find_clusters_from_adjacency} uses breadth-first search to identify the connected components
of the corresponding adjacency graph of the centroid differences vectors.
}
