% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{GKernel_weights}
\alias{GKernel_weights}
\title{\code{GKernel_weights} computes the Gaussian kernel weights, 
using the distance matrix as the input.}
\usage{
GKernel_weights(phi, distance)
}
\arguments{
\item{phi}{The parameter in the Gaussian kernel}

\item{distance}{The distance matrix, with dimension n*n, 
where n is the row count of the data matrix. The distance matrix can be generated 
by function \code{distance_matrix} from the data matrix.}
}
\description{
\code{GKernel_weights} computes the Gaussian kernel weights, 
using the distance matrix as the input.
}
