\name{Rcvxcluster-package}
\alias{Rcvxcluster-package}
\alias{Rcvxcluster}
\docType{package}
\title{
Robust Convex Clustering
}
\packageTitle{Rcvxcluster}
\description{
Implement a robust convex clustering method, with Huber loss and fusion penalty, of which weights can be generated by a new approach that is more robust to outliers. See reference. 
}
\packageDescription{Rcvxcluster}
\details{
}
\packageDESCRIPTION{Rcvxcluster}
\packageIndices{Rcvxcluster}
\author{}
\packageAuthor{Rcvxcluster}
Maintainer: \packageMaintainer{Rcvxcluster}
\references{
Liu, Chenyu & Sun, Qiang & Tan, Kean. (2019). Robust convex clustering: How does fusion penalty enhance robustness?. https://arxiv.org/abs/1906.09581v2
}
\keyword{ package }
\seealso{}
\examples{}
