% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{robust_weights}
\alias{robust_weights}
\title{Compute weights}
\usage{
robust_weights(X, delta, zeta)
}
\arguments{
\item{X}{The data matrix to be clustered. The rows are observations, and the columns 
are features.}

\item{delta}{The nonnegative parameter that controls the scale of robust weights 
when there is outlier(s) in the data.}

\item{zeta}{The nonnegative parameter that controls the scale of robust weights.}
}
\value{
A vector \cite{wt} of weights for robust convex clustering.
}
\description{
\code{robust_weights} computes the robust weights given a data matrix \code{X}, 
a scale parameter \code{zeta} and a parameter that controls the weights \code{delta}.
Namely, the lth weight \code{w[l]} is given by 
\deqn{
w[l] = exp(-\zeta{\sum_{j\in D_{1}}(X_{i'j}-X_{ij})^2+\sum_{j\in D_{2}}{\delta}^2})
, where the lth pair of nodes is (\code{i},\code{i'})
and \code{D1={j:|X_{ij}-X_{i'j}|<delta}}, \code{D2={j:|X_{ij}-X_{i'j}|>delta}}.
}
}
\author{
Chenyu Liu, Qiang Sun, Kean Ming Tan
}
