% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustcvxclust.R
\name{robustcvxclust}
\alias{robustcvxclust}
\title{Robust Convex Clustering}
\usage{
robustcvxclust(
  X,
  W = NULL,
  V = NULL,
  Y = NULL,
  Z = NULL,
  max_iter = 1e+05,
  rho,
  tau,
  lambda,
  wt,
  tol_abs = 1e-05
)
}
\arguments{
\item{X}{The n-by-p data matrix whose rows are being clustered.}

\item{W}{The n-by-p matrix of Lagrange multipliers.}

\item{V}{The centroid difference matrix.}

\item{Y}{The nK-by-p matrix of Lagrange multipliers.}

\item{Z}{The n-by-p matrix of Lagrange multipliers.}

\item{max_iter}{The maximum number of iterations. The default value is 1e5.}

\item{rho}{Augmented Lagrangian penalty parameter.}

\item{tau}{The robustification parameter in huber loss.}

\item{lambda}{The regularization parameter controlling the amount of shrinkage.}

\item{wt}{A vector of nK positive weights.}

\item{tol_abs}{The convergence tolerance. The default value is 1e-05.}
}
\value{
\code{U} The centroid matrix.

\code{W} The centroid matrix.

\code{V} The centroid difference matrix.

\code{Y} The Lagrange multiplier matrix.

\code{Z} The Lagrange multiplier matrix.

\code{iter} The number of iterations taken.

\code{tol} The residual tolerances.
}
\description{
\code{robustcvxclust} performs robust convex clustering using ADMM. This is an R wrapper function around C code.
Dimensions of various arguments are as follows:
\itemize{
\item{n is the number of data observations}
\item{p is the number of features}
\item{nK is the number non-zero weights.}
}
}
\author{
Chenyu Liu, Qiang Sun, Kean Ming Tan
}
