% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustcvxclust_path.R
\name{robustcvxclust_path}
\alias{robustcvxclust_path}
\title{Robust Convex Clustering}
\usage{
robustcvxclust_path(X, wt, lambda, rho, tau, tol_abs = 1e-05, max_iter = 1e+05)
}
\arguments{
\item{X}{The data matrix to be clustered. The rows are the observations, and the columns are the features.}

\item{wt}{A vector of nonnegative weights. The ith entry \code{w[i]} denotes the weight used between the ith pair of centroids.}

\item{lambda}{A sequence of regularization parameters.}

\item{rho}{Augmented Lagrangian penalty parameter.}

\item{tau}{The robustification parameter in huber loss.}

\item{tol_abs}{The convergence tolerance (absolute).}

\item{max_iter}{The maximum number of iterations.}
}
\value{
\code{U} A list of centroid matrices.

\code{W} A list of centroid matrices.

\code{V} A list of centroid difference matrices.

\code{Y} A list of Lagrange multiplier matrices.

\code{Z} A list of Lagrange multiplier matrices.
}
\description{
\code{robustcvxclust_path} estimates the robust convex clustering path using ADMM.
Required inputs include a data matrix \code{X} (rows are the observations; columns are the features), a vector of weights \code{wt}, and a sequence of regularization parameters \code{lambda}. The penalty norm supported is 2-norm.
}
\author{
Chenyu Liu, Qiang Sun, Kean Ming Tan
}
