\name{Soft}
\alias{Soft}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Soft-thresholding Operator
}
\description{
Soft(a,b)=sign(a)max(|a|-b,0))
}
\usage{
Soft(a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{The variable a.}
  \item{b}{The threholding b.}
}
\details{}
\value{
}
\references{
Liu, Chenyu & Sun, Qiang & Tan, Kean. (2019). Robust convex clustering: How does fusion penalty enhance robustness?. https://arxiv.org/abs/1906.09581v2
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
