% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{uni_weights}
\alias{uni_weights}
\title{\code{uni_weights} computes the uniform weights.}
\usage{
uni_weights(n, p)
}
\arguments{
\item{n}{The row count, or amount of data points, of the data matrix.}

\item{p}{The column count, or amount of features, of the data matrix.}
}
\description{
\code{uni_weights} computes the uniform weights.
}
